Refactor project structure
core.structures.* includes data structures and types relevant to the
implementation of the game of life. This is distinct from
datastructures.* in that the latter provides functionality to the core
structures and the core structures provide functionality to the Game.

The game logic, not including invocation information is stored in
core.logic.* this includes all mechanisms for managing the board,
applying the rules (which ultimately change the board) and so on.

Invocation information (including main()) is stored in main.Main.java .
Main will setup the input and output, making use of ui.* for receiving
instructions to change the game, and then transmitting the results
generated by core.logic.* .

Interfaces for all packages x (if applicable) are stored in
x.interfaces.* .