Fix ImmutableSet usage of internal set
Collections.emptySet() is immutable, which is not what we need for the
immutable set type. The immutable set needs to store items, but cannot
do so if it's internal set is itself immutable.

Using a hashset, which is now acceptable because of the hashCode
additions to the basic data structures