Refactor internal ImmutableSet from Set and Map to a Trie like structure
The set and map, although lazy, are not much more efficient than an
immutable set that copies it's elements on each action. This is because
we copy the map of changes each time, and necessarily copy the permanent
elements. While the changes map may be small we are not benefiting much
from the immutability, although the laziness does reduce the cost
somewhat.

This new implementation is similar to a trie, to be explained in the
upcoming documentation, which should almost remove the cost of copying
elements on derivative sets.