Add AssetLoader framework
The asset loaders are Main's point of contact with Displayables.
AssetLoaders create Displayables (throwing any necessary exceptions on
load), and the Displayables are then used as is with the assurance that
they will not throw exceptions

AssetLoaders have an opportunity to display messages to the user via
stderr. This is done through the error message in the exception thrown
by the AssetLoader. This is likely to change in future with the addition
of an interactive game of life

The AssetLoaders have one obvious limitation: all assets need to be
loaded at the start; if this becomes a problem we can deal with it then