Change definition of Displayable
Displayable now:
Extends GenericObserver<Board> so all Displayables receive updates on
new boards
Defines run() which is called after main has made all the calls it needs
to. run() will not be called if the Displayable could not be initialised
for any reason - the reason will then be printed to stderr as per the
AssetLoader interface

Displayables are expected to setup their own GenericObservable for the
running of the game - GameManager is provided for this purpose. The
alternative is for main to create a GameManager for the Displayables,
almost all of which would throw it out to gather their own input from
the user (via a GUI, for example) to create a new GameManager and start
it when the user is ready which is unlikely to be when run is called