Add CLI argument parsing to GraphicalAssetLoader. Add GameManager to GUI
The asset loader attempts to load a provided alive or dead image. If
they are not provided defaults are used: src/ui/gui/assets/alive.png and
[...]/dead.png respectively.

The GUI now uses the GameManager to provide updates to the board and to
execute the game rules. Currently the board and speed are hardcoded, but
upcoming features will include setting the speed and drawing on board